Add signature to check for recently common CNI configuration error
A lot of recent production installation failures seem to be caused by an
error where kubelet complains about CNI plugin configuration.

This new signature looks for this repetitive error in the kubelet
journal of one of the control plane nodes, and if it finds it more
than threshold (1000) times, it creates the signature.

The journal logs are found in the installer-gather bootstrap log bundle,
under the control-plane masters-gather directory.