Remove SMART error message for disk inspection
As part of the agent discovery we are collecting SMART data about the
disks present on the host. As there is a lot of drives and controllers
that do not return correct data, we are often encountering non-zero exit
code from smartctl command. In most of the cases the message, instead of
providing a real value, confuses engineers investigating failures of the
cluster installation.

This PR removes the part responsible for printing the error message
returned by smartctl. This means that only parseable fields will be
attached to the AITRIAGE ticket, but the raw error from the command will
not be printed.

Contributes-to: MGMT-7388