Handle error in smartctl command execution
This commit adds a clarification when the smartctl command failed
because of the internal error rather than because of the disk failure.
It leverages the bit 0 of the exit status bitmask which is defined as
"command line did not parse".

Ref.: https://www.smartmontools.org/static/doxygen/smartctl_8h_source.html