Allow specifying more domains for automatic AITRIAGE closure (#166)
This PR changes the logic of the automatic triage tickets to allow
specifying a list of domains from which the tickets should be
automatically closed.

This is useful in scenarios when e.g. we know about the user who is
developing a new feature that we know is prone to fail in SaaS. In order
to remove the noise, those tickets are better to be automatically
closed.