NO-ISSUE: retrying when fetching comments
We sometimes get 401 Unauthorized status code from Jira service when we
try to fetch comments related to an issue. This change will make 3 tries
for this action (which does not have any side-effects as it's a GET
request) with delay between requests of 2 seconds (so that if the SSO
service is temporarily unavailable we'll try again afterwards).