Added new OSInstallationTime signature for hosts that took a long time to write RHCOS to disk
For each host in the cluster, finds the:

```python
r'.*reached installation stage Writing image to disk$'
```

and

```python
r'.*reached installation stage Writing image to disk: 100%$'
```

events, takes the difference between their timestamps, and if it goes
over the threshold (300 seconds), it adds a warning about that
host which says how long writing the image to disk took.

If there's at-least one slow host, a table will be generated for all
hosts, showing the duration of all of them, even if they are below the
threshold.