Better handling of events file containing events from multiple installation attempts
When we collect triage logs for a single ticket, we also collect all
events from previous installation attempts for that cluster. However,
multiple installation of the same cluster get separate Jira tickets.

This commit adjusts our triage signatures to only look at events from
the final installation attempt of this cluster for which this ticket was
created.

I've also added a new signature that warns triagers that the events file
for this ticket contains events from multiple installation attempts, and
tells them the timestamp in which the first event for the last
installation attempt (to which the ticket is relevant) happened.

The way we tell apart events from different installation attempts is by
partitioning the event list according to `"cluster_installation_reset"`
events.