Improve `SNOMachineCidrSignature` to handle multiple default routes
It seems like OCP will hit etcd certificate errors even when there
are multiple default routes where one of them doesn't match the
machine_cidr.

We saw this happen in https://issues.redhat.com/browse/AITRIAGE-3541
and the signature `SNOMachineCidrSignature` wasn't triggered for that
ticket seemingly because it had multiple default routes, with the first
one matching the machine CIDR and the second one not matching.