Ticket labels should only be re-applied if the comment was changed (#280)
It only makes sense to re-evaluate ErrorSignature labels for a ticket if the
signature comment didn't already exist or if it did exist and the user
explicitly asked to to re-evaluate.

This is because by default we don't want to re-evaluate signatures for
existing tickets that already had that signature evaluated for them, we
only want to do it when should_reevaluate is true (when the user passes
the `--update` flag).