MGMT-11688: Skip bump of 4.11 rhcos images (#285)
Currently the bump-ocp job is unable to fetch rhcos image for 4.11:
```
requests.exceptions.HTTPError: 404 Client Error: Not Found for url:
https://mirror.openshift.com/pub/openshift-v4/x86_64/dependencies/rhcos/4.11/4.11.0/rhcos-4.11.0-x86_64-live.x86_64.iso
```

Root-cause for this seems to be due to
https://issues.redhat.com/browse/ART-4439, meaning we'll have to skip it
for now until a fix is delivered.

This change introduces two separate skip lists for OCP and RHCOS images,
enabling us to just skip updating 4.11 rhcos image (and continue bumping
the OCP images).

Also, it renames the variables because the y-part in x.y.z is actually
called minor.