MGMT-11987: fix 4.12 versions getting overridden with 4.11 versions (#300)
Apparently the candidate-4.12 channel is pre-filled with 4.11 versions
so that people are able to upgrade from 4.11 to 4.12rc immediately:
```
$ curl -s
"https://api.openshift.com/api/upgrades_info/v1/graph?channel=candidate-4.12"
| jq -r '.nodes[].version'
4.11.0-rc.1
4.11.0-fc.3
4.11.0-rc.3
4.11.0-rc.0
4.11.0
...
```

This means with current logic we might override versions in cases there
is no actual target version (4.x) in a ``*-4.x`` channel (but only
``4.x-1``) versions.

This change brings back the automatic handling of 4.12 versions, and
fixes the logic to only include versions that match the minor-version
that is searched for.