TagAnalysis signature - inform user about known/unknown tags
Clusters created with the SaaS Assisted Installer can be given arbitrary
tags.

The UI automatically gives all clusters the ui_ocm tag, and aicli always
gives cluster the aicli tag.

The TagAnalysis maintains a list of tags with a known meaning and
explains that meaning in a triage signature, to give the people
doing triage more context about how this cluster was installed.

More tags should be added in the future once they are known