workaround to must-gather sig (#331)
currently the signature show up in cases that we know that the
must-gather can't be collected.
With this fix triage tickets will get this signature only if must-gather
logs should've been collected - the bootstrap node went to reboot.
In case of SNO the signature will be added only if we the controller
start running post reboot.