MGMT-13704: Adapt signatures to get host logs both with new and old paths
While we transition to the new logs path structure (see MGMT-13454), signatures
should search for logs both in the old and the new location (in case the
former failed).

Once MGMT-13454 lands in production and is confirmed to be working as
expected, we will remove the old path from the signatures code base.