MGMT-13752: Fix description not filled in (#354)
Back in
https://github.com/openshift-assisted/assisted-installer-deployment/pull/353
I acted on the assumption that filling-in the description right at
ticket creation is not required, as later on we're activating the
``FailureDescription`` signature (which is what's filling-in the ticket
description). It turns out that the code reads logs URL from the
description, and that the signature doesn't apply by default.

This changes the code to be a bit more coherent around the
source-of-truth for getting logs base URL, as well as making the
signatures less dependent on each other.