MGMT-13667: Signature to detect user login to the cluster hosts.
During installation, if the user logs into one of the cluster hosts this might be something we need to know about.
Whereas we acknowledge that this does not prove that the user made a breaking change, it is an indicator that we should look closer for such things.

So we are adding this signature to detect a manual login to any of the cluster hosts.