MGMT-13401: store cluster logs as Jira attachments (#338)
We plan moving from the logs-collector server to a more maintainable and secure solution, being Jira attachments.

This change first creates Jira tickets as placeholder for each installation attempt (it no existing ticket can be found), archive and split the files (because Jira attachments have a size limit of 50MB) and continue with the rest of the process.

We shouldn't have lots of cases where we split the files (only 0.02% of the logs are more than 50MB), and for now this does not change the further processing through logs-collector server. Later on we'll create a method for reading Jira attachments and editing the signatures based on those.