Split long triage status reports
Slack is automatically splitting long messages (when it's more than 4040
characters long). Affects of this are not taht good, because it causes a
malformed markdown format.

This change splits the report to several messages (one message after another.
Unfortunately it's not seem possible to do in one thread when using
webhooks and the webhook's client), when the table has more than 7 rows.
If there are 7 rows or less, it just fits it in one single message (as
it's doing right now).

The change also leverages use of ``slack_sdk`` library, which should verify
enable a better handling of requests and responses.